Protocol classes generation (#232)
* started poc

* wip

* almost done

* save

* it's alive

* it works with commands

* starting events

* [wip] add protocol

* formatting

* completed generation

* prefix events with namespace

* cleanup

* formatting

* format

* rename file

* formatting

* format

* internal models

* cr

* it compiles

* format

* format

* generator

* format

* some progress

* wip

* test not done yet

* better chromium events

* better generator

* format

* better generator

* skip wip test

* ignore schromium

* cr

* cleanup

* better generated code

* working

* cr

* style

* cleanup

* nullable connection response properties

* cr

* fix warnings

* cleanup

+++ src/PlaywrightSharp.Abstractions/GeolocationOption.cs
@@ -12,19 +12,19 @@ namespace PlaywrightSharp
         /// Latitude between -90 and 90.
         /// </summary>
         /// <value>The latitude.</value>
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
 
         /// <summary>
         /// Longitude between -180 and 180.
         /// </summary>
         /// <value>The longitude.</value>
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
         /// <summary>
         /// Optional non-negative accuracy value.
         /// </summary>
         /// <value>The accuracy.</value>
-        public decimal Accuracy { get; set; }
+        public double Accuracy { get; set; }
 
         /// <inheritdoc cref="IEquatable{T}"/>
         public bool Equals(GeolocationOption other)
